week4/project-5-agent-api/requirements.txt
fastapi
uvicorn[standard]
langchain-anthropic
langgraph
langchain-core
python-dotenv
requests
